+++ examples/simulated/plot_mean_median_comparison.py
@@ -16,9 +16,8 @@ import matplotlib.pyplot as plt
 
 from sklearn.datasets import make_blobs
 from pyriemann.datasets import make_outliers
-from pyriemann.utils.mean import (
-    mean_euclid, mean_riemann, median_euclid, median_riemann
-)
+from pyriemann.utils.mean import mean_euclid, mean_riemann
+from pyriemann.utils.median import median_euclid, median_riemann
 from pyriemann.clustering import Potato
 
 rs = np.random.RandomState(17)
@@ -74,10 +73,10 @@ plt.show()
 
 n_channels, n_inliers, n_outliers = 2, 16, 6
 
-Xin = make_outliers(
-    n_inliers, 0.2*np.eye(n_channels), 0.5, outlier_coeff=1, random_state=rs)
+Cin = 0.2 * np.eye(n_channels)
+Xin = make_outliers(n_inliers, Cin, 0.5, outlier_coeff=1, random_state=rs)
 Xout = make_outliers(
-    n_outliers, 4*np.eye(n_channels), 0.5, outlier_coeff=1, random_state=rs)
+    n_outliers, 4 * np.eye(n_channels), 0.5, outlier_coeff=1, random_state=rs)
 X = np.concatenate([Xin, Xout])
 
 C_emean = mean_euclid(X)
@@ -88,6 +87,7 @@ C_rmed = median_riemann(X)
 fig2 = plt.figure(figsize=(7, 7))
 fig2.suptitle("Means and medians for 2x2 SPD matrices", fontsize=16)
 ax2 = plt.subplot(111, projection='3d')
+ax2.scatter(1, 0, 1, c="k", marker="+", s=50, label='Identity')
 ax2.scatter(Xin[:, 0, 0], Xin[:, 0, 1], Xin[:, 1, 1], c="C0", edgecolors="k",
             label='Inliers')
 ax2.scatter(Xout[:, 0, 0], Xout[:, 0, 1], Xout[:, 1, 1], c="C1",
@@ -100,7 +100,6 @@ ax2.scatter(C_emed[0, 0], C_emed[0, 1], C_emed[1, 1], c="r", marker="s",
             label='Euclidean median')
 ax2.scatter(C_rmed[0, 0], C_rmed[0, 1], C_rmed[1, 1], c="m", marker="s",
             label='Riemannian median')
-ax2.scatter(1, 0, 1, c="k", marker="+", s=50, label='Identity')
 ax2.legend(loc='center left', bbox_to_anchor=(0.7, 0.6))
 plt.show()
 
@@ -109,16 +108,19 @@ plt.show()
 # Photo finish
 # ------------
 #
-# Specific zoom on means and medians, with a surprise guest.
+# Specific zoom on means and medians.
 #
-# Riemannian potato is fitted with an iterative outlier removal, providing a
-# robust mean [4]_.
+# Surprise guest: Riemannian potato is fitted with an offline iterative outlier
+# removal, providing a robust mean [4]_.
 
-C_rp = Potato(metric='riemann', threshold=1.5).fit(X)._mdm.covmeans_[0]
+C_rp = Potato(metric='riemann', threshold=1.5).fit(X).covmean_
 
 fig3 = plt.figure(figsize=(7, 7))
 fig3.suptitle("Means and medians for 2x2 SPD matrices\nZoom", fontsize=16)
 ax3 = plt.subplot(111, projection='3d')
+ax3.scatter(1, 0, 1, c="k", marker="+", s=50, label='Identity')
+ax3.scatter(Cin[0, 0], Cin[0, 1], Cin[1, 1], c="C0", edgecolors="k", s=50,
+            label='Center of inliers')
 ax3.scatter(C_emean[0, 0], C_emean[0, 1], C_emean[1, 1], c="r", marker="x",
             label='Euclidean mean')
 ax3.scatter(C_rmean[0, 0], C_rmean[0, 1], C_rmean[1, 1], c="m", marker="x",
@@ -127,9 +129,8 @@ ax3.scatter(C_emed[0, 0], C_emed[0, 1], C_emed[1, 1], c="r", marker="s",
             label='Euclidean median')
 ax3.scatter(C_rmed[0, 0], C_rmed[0, 1], C_rmed[1, 1], c="m", marker="s",
             label='Riemannian median')
-ax3.scatter(C_rp[0, 0], C_rp[0, 1], C_rp[1, 1], c="b", marker="*",
-            label='Riemannian potato')
-ax3.scatter(1, 0, 1, c="k", marker="+", s=50, label='Identity')
+ax3.scatter(C_rp[0, 0], C_rp[0, 1], C_rp[1, 1], c="g", marker="*",
+            label='Center of\nRiemannian potato')
 ax3.legend(loc='center left', bbox_to_anchor=(0.7, 0.5))
 plt.show()