+++ doc/whatsnew.rst
@@ -21,7 +21,7 @@ v0.3.1.dev
 
 - Add parameter ``sampling_method`` to :func:`pyriemann.datasets.sample_gaussian_spd`, with ``rejection`` accelerating 2x2 matrices generation. :pr:`198` by :user:`Artim436`
 
-- Add geometric medians for Euclidean and Riemannian metrics: :func:`pyriemann.utils.mean.median_euclid` and :func:`pyriemann.utils.mean.median_riemann`. :pr:`200` by :user:`qbarthelemy`
+- Add geometric medians for Euclidean and Riemannian metrics: :func:`pyriemann.utils.median.median_euclid` and :func:`pyriemann.utils.median.median_riemann`. :pr:`200` by :user:`qbarthelemy`
 
 v0.3 (July 2022)
 ----------------